extract signing methods as standalone service
