+++ src/Digiflow.Payment/Interfaces/ISignService.cs
@@ -1,0 +1,6 @@
+namespace Digiflow.Payment.Interfaces;
+
+public interface ISignService
+{
+	string Sign(IBaseQueryParam param);
+}
